+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # MiniReel Player
 
 ## Beta12.1 (June 16, 2014)
+* *[Beta12.1.rc3]*
+  * If a companion is shown at the same time as a video ad, and the ad
+    has already been viewed, the ad card is skipped
+
 * *[Beta12.1.rc2]*
   * [FIX]: Clean up ad view in lightbox without display ads
   * [FIX]: Added adtech secure property to ad calls